Fix: 64-bit end addresses in symbol processing

+++ etrace.c
@@ -282,7 +282,7 @@ void etrace_process_exec(struct etracer *t, enum cov_format cov_fmt)
 			if (sym) {
 				uint64_t addr = start;
 				while (sym && addr < end) {
-					uint32_t tend = end;
+					uint64_t tend = end;
 
 					if (tend > (sym->addr + sym->size)) {
 						tend = sym->addr + sym->size;
